--- src/components/Carousel/ImageSlider/styles.css
@@ -1,23 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
-
-.bg {
-  position: absolute;
-  top: 100px;
-  left: 225px;
-  width: 800px;
-  height: 300px;
-  background-size: cover;
-  background-position: center;
-  will-change: opacity;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.bg h1 {
-  font-size: 120px;
-  color: #fff;
-  text-shadow: 0px 6px 6px rgba(0, 0, 0, 0.4);
-}
